- rand states inited before optix launch - moved sample loop to host code, - optixLaunch called for each sample - pixels avged after the loop

+++ src/renderer.h
@@ -14,7 +14,7 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-	CHIZEN_RESULT renderer_render_gltf(const size_t render_width, const size_t render_height, const char* gltf_path, exr_pass* passes, const size_t passes_pixels_count);
+	CHIZEN_RESULT renderer_render_gltf(const uint32_t render_width, const uint32_t render_height, const char* gltf_path, exr_pass* passes, const uint32_t passes_pixels_count);
 #ifdef __cplusplus
 }
 #endif
